+++ src/parman/tests/test_demos.py
@@ -18,6 +18,8 @@
 #
 # --
 """Run the demos to verify that they run without errors."""
+
+
 import hashlib
 import importlib.resources
 import os
@@ -28,18 +30,36 @@ from pathlib import Path
 
 import pytest
 
-JOBDEMO_RELPATHS = [
-    Path("jobdemo.py"),
-    Path("templates/boot/jobinfo.py"),
-    Path("templates/boot/run"),
-    Path("templates/compute/jobinfo.py"),
-    Path("templates/compute/run"),
-    Path("templates/sample/jobinfo.py"),
-    Path("templates/sample/run"),
-    Path("templates/train/helper.py"),
-    Path("templates/train/jobinfo.py"),
-    Path("templates/train/run"),
-]
+
+def test_linreg(tmppath: Path):
+    run_script(["python3", "linreg.py"], Path("demos/demc"), [Path("linreg.py")], tmppath)
+
+
+def test_naivemc(tmppath: Path):
+    run_script(
+        ["python3", "naivemc.py"],
+        Path("demos/demc"),
+        [Path("linreg.py"), Path("naivemc.py")],
+        tmppath,
+    )
+
+
+def test_demc_serial(tmppath: Path):
+    run_script(
+        ["python3", "demc.py"],
+        Path("demos/demc"),
+        [Path("linreg.py"), Path("naivemc.py"), Path("demc.py")],
+        tmppath,
+    )
+
+
+def test_demc_parman(tmppath: Path):
+    run_script(
+        ["python3", "demc.py", "--parman"],
+        Path("demos/demc"),
+        [Path("linreg.py"), Path("naivemc.py"), Path("demc.py")],
+        tmppath,
+    )
 
 
 @pytest.mark.parametrize(
@@ -63,23 +83,74 @@ JOBDEMO_RELPATHS = [
     ],
 )
 def test_jobdemo(framework: str, schedule: bool, in_place: bool, tmppath: Path):
-    # Put the files of the jobdemo in place
-    for relpath in JOBDEMO_RELPATHS:
-        dn_dst = tmppath / relpath.parent
-        dn_dst.mkdir(parents=True, exist_ok=True)
-        shutil.copy(Path("demos/jobdemo-al") / relpath, tmppath / relpath)
-
-    # Run the jobdemo in a subprocess. (runpy does not work with parsl.)
     args = ["python3", "jobdemo.py", framework, "--pause=0"]
     if schedule:
         args.append("-s")
     if in_place:
         args.append("-i")
+
+    relpaths = [
+        Path("jobdemo.py"),
+        Path("templates/boot/jobinfo.py"),
+        Path("templates/boot/run"),
+        Path("templates/compute/jobinfo.py"),
+        Path("templates/compute/run"),
+        Path("templates/sample/jobinfo.py"),
+        Path("templates/sample/run"),
+        Path("templates/train/helper.py"),
+        Path("templates/train/jobinfo.py"),
+        Path("templates/train/run"),
+    ]
+
+    # Run the jobdemo in a subprocess. (runpy does not work with parsl.)
+    run_script(args, Path("demos/jobdemo-al"), relpaths, tmppath)
+    if framework != "dry":
+        check_files(tmppath, "jobdemo-al-results.sha256")
+
+
+def test_plastic_ibuprofen(tmppath: Path):
+    run_script(
+        [
+            "python3",
+            "plastic.py",
+            "ibuprofen.xyz",
+            "ibuprofen.traj",
+            "ibuprofen_final.xyz",
+            "--steps=100",
+        ],
+        Path("demos/inspiration/plastic"),
+        [Path("plastic.py"), Path("ibuprofen.xyz")],
+        tmppath,
+    )
+
+
+def test_plastic_alumina(tmppath: Path):
+    run_script(
+        [
+            "python3",
+            "plastic.py",
+            "alumina.json",
+            "alumina.traj",
+            "alumina_final.xyz",
+            "--steps=100",
+        ],
+        Path("demos/inspiration/plastic"),
+        [Path("plastic.py"), Path("alumina.json")],
+        tmppath,
+    )
+
+
+def run_script(args: list[str], root: Path, relpaths: list[Path], tmppath: Path, nrepeat: int = 1):
+    """Run a script with auxiliary files in a temporary directory."""
+    # Put the files in place
+    for relpath in relpaths:
+        dn_dst = tmppath / relpath.parent
+        dn_dst.mkdir(parents=True, exist_ok=True)
+        shutil.copy(root / relpath, tmppath / relpath)
+    # Execute the script
     env = os.environ | {"PYTHONPATH": Path("src/").absolute()}
-    for _ in range(2):
+    for _ in range(nrepeat):
         subprocess.run(args, check=True, cwd=tmppath, env=env)
-        if framework != "dry":
-            check_files(tmppath, "jobdemo-al-results.sha256")
 
 
 def check_files(root, fn_sha):