+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- A pure Python usage example (Differential Evolution Monte Carlo).
+- The `parman-sbatch-wait` script to facilitate running workflow jobs on a Slurm cluster.
+
 ## [0.1.0] - 2023-05-02
 
 Initial public release. See `README.md` for some preliminary details.